[metal] Move platform/metal to backends/metal
