+++ taichi/backends/metal/kernel_util.cpp
@@ -1,4 +1,4 @@
-#include "taichi/platform/metal/metal_kernel_util.h"
+#include "taichi/backends/metal/kernel_util.h"
 
 #define TI_RUNTIME_HOST
 #include "taichi/runtime/llvm/context.h"
